# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ccce-atlas API",
        "description": "Civic geospatial platform API for Corpus Christi / Nueces County, Texas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/parcels/{parcel_id}/geometry": {
            "get": {
                "tags": [
                    "parcels"
                ],
                "summary": "Get Parcel Geometry",
                "description": "Get just the geometry for a parcel (lightweight endpoint for boundary highlighting).\n\nArgs:\n    parcel_id: Unique parcel identifier\n\nReturns:\n    Parcel geometry as GeoJSON",
                "operationId": "get_parcel_geometry_api_v1_parcels__parcel_id__geometry_get",
                "parameters": [
                    {
                        "name": "parcel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Parcel Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParcelGeometry"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/parcels/tiles/{z}/{x}/{y}.json": {
            "get": {
                "tags": [
                    "parcels"
                ],
                "summary": "Get Parcel Tile",
                "description": "Get parcel tile as GeoJSON for a given z/x/y coordinate.\n\nThis endpoint serves vector tiles in GeoJSON format for efficient\nrendering of large parcel datasets in Cesium.\n\nArgs:\n    z: Zoom level (0-22)\n    x: Tile X coordinate\n    y: Tile Y coordinate\n\nReturns:\n    GeoJSON FeatureCollection with parcels in the tile",
                "operationId": "get_parcel_tile_api_v1_parcels_tiles__z___x___y__json_get",
                "parameters": [
                    {
                        "name": "z",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Z"
                        }
                    },
                    {
                        "name": "x",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "X"
                        }
                    },
                    {
                        "name": "y",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Y"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ParcelGeometry": {
                "properties": {
                    "parcel_id": {
                        "type": "string",
                        "title": "Parcel Id"
                    },
                    "geometry": {
                        "title": "Geometry"
                    }
                },
                "type": "object",
                "required": [
                    "parcel_id",
                    "geometry"
                ],
                "title": "ParcelGeometry",
                "description": "Just the geometry for a parcel (lightweight response for boundary highlighting)."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}